Explicit cast to (uint8_t)
This variable comes from
    namelen = strlen(name)
where `name` is an argument passed to netpacket_register_cha_response().  This
function is only called in
    keygen/keygen_network.c
where the value `C->name` is passed.  That value comes from
    keygen/keygen.c
but it is checked in
    keygen/keygen_actual.c
which ensures that
    0 < strlen(C->name) <= 255

+++ lib/netpacket/netpacket_register.c
@@ -64,7 +64,7 @@ netpacket_register_cha_response(NETPACKET_CONNECTION * NPC,
 
 	/* Construct challenge response packet. */
 	memcpy(packetbuf, keys, 96);
-	packetbuf[96] = namelen;
+	packetbuf[96] = (uint8_t)namelen;
 	memcpy(packetbuf + 97, name, namelen);
 
 	/* Append hmac. */
